Add latest changes to case
Case has now holes for scart, HDMI, RCA, and VGA in place. There is also
a holder for HDMI adapter.

The model has an assembly where the PCB, case, and HDMI holder are all
positioned in their real positions.

Models for case and HDMI holder are outputted as STL when ran.